Split user CRUD tests into multiple files
